% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeoCoding.R
\name{getGeoCoding}
\alias{getGeoCoding}
\title{Get longitude and latitude through Baidu map API}
\usage{
getGeoCoding(
  ak = NULL,
  address = NULL,
  city = NULL,
  ret_coordtype = "gcj02ll",
  sn = NULL,
  output = "json"
)
}
\arguments{
\item{ak}{user must register their ak (key) before use.}

\item{address}{the address to parse}

\item{city}{the city of address}

\item{ret_coordtype}{the type of coordinate}

\item{sn}{not necessary}

\item{output}{must by json}
}
\value{
Longitude and latitude of address
}
\description{
Get longitude and latitude through Baidu map API
}
\details{
[Detail document reference](https://lbsyun.baidu.com/index.php?title=webapi/guide/webservice-geocoding)
}
\examples{
getGeoCoding(ak = '92RLPSx3RqhlaNGjGTdzL2f2iqHmgl4f', address = 'beijing')

}
