% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeoCoding.R
\name{getGeoCoding}
\alias{getGeoCoding}
\title{Geocoder of Baidu map API, https://lbsyun.baidu.com/index.php?title=webapi/guide/webservice-geocoding}
\usage{
getGeoCoding(
  ak = NULL,
  address = NULL,
  city = NULL,
  ret_coordtype = "gcj02ll",
  sn = NULL,
  output = "json"
)
}
\arguments{
\item{ak}{, 用户申请注册的key，自v2开始参数修改为“ak”}

\item{address}{, 待解析的地址。最多支持84个字节。可以输入两种样式的值，分别是： 1、标准的结构化地址信息，如北京市海淀区上地十街十号 【推荐，地址结构越完整，解析精度越高】2、支持“*路与*路交叉口”描述方式，如北一环路和阜阳路的交叉路口}

\item{city}{,地址所在的城市名。用于指定上述地址所在的城市，当多个城市都有上述地址时，该参数起到过滤作用，但不限制坐标召回城市。}

\item{output}{}
}
\value{

}
\description{
Geocoder of Baidu map API, https://lbsyun.baidu.com/index.php?title=webapi/guide/webservice-geocoding
}
